requirements.txt
  livekit-agents[deepgram,elevenlabs,openai,cartesia,silero,turn-detector]~=1.0
  livekit-plugins-noise-cancellation~=0.2
  python-dotenv